requirements.txt

asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.46
botocore==1.27.46
dj-database-url==1.0.0
Django==3.2.9
django-countries==7.3.2
django-country==0.1.0
django-decouple==2.1
django-filter==22.1
django-heroku==0.3.1
django-storages==1.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
importlib-metadata==4.12.0
jmespath==1.0.1
Markdown==3.4.1
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
whitenoise==6.2.0
zipp==3.8.1
django-cors-headers==3.13.0